# docs/web-control-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChatWire local moderator API",
    "version": "1.0.0",
    "description": "Private moderator dashboard API. Issue browser access via the Discord /web command. No public endpoint issues grants. See web-control-setup.md for deployment and pending settings semantics."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "summary": "Latest jobs grouped by instance plus host batches",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job}": {
      "get": {
        "summary": "Read a host batch job",
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/batches/preview": {
      "post": {
        "summary": "Preview explicit target list",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain from GET /api/v1/me; Origin must equal the configured public origin."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Batch"
              }
            }
          }
        }
      }
    },
    "/api/v1/batches": {
      "post": {
        "summary": "Submit confirmed RCON or config-reload batch",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain from GET /api/v1/me; Origin must equal the configured public origin."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "description": "Reuse the same key and body after a lost response; do not generate a replacement key."
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Structured error; status describes authentication, validation, conflict, or availability failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Batch"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-chatwire"
      }
    },
    "schemas": {
      "Actor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "admin": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "admin"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "server_id": {
            "type": "string"
          },
          "actor": {
            "$ref": "#/components/schemas/Actor"
          },
          "action": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "waiting_for_empty",
              "waiting_for_restart",
              "succeeded",
              "failed",
              "unknown",
              "interrupted"
            ]
          },
          "phase": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "result": {},
          "error": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "server_id",
          "action",
          "state"
        ]
      },
...
      "ActionParameters": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "reason": {
            "type": "string"
          },
          "save_id": {
            "type": "string"
          },
          "exchange_string": {
            "type": "string"
          },
          "command": {
            "type": "string"
          },
          "operation": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "player": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "confirmation_token": {
            "type": "string"
          },
          "when_empty": {
            "type": "boolean"
          },
          "force": {
            "type": "boolean"
          },
          "start": {
            "type": "boolean"
          },
          "mods": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "upload_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "level": {
            "type": "integer",
            "enum": [
              -255,
              -1,
              0,
              1,
              2,
              3,
              255
            ]
          }
        }
      },
      "Batch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "rcon",
              "config-reload"
            ]
          },
          "servers": {
            "type": "array",
            "minItems": 1,
            "maxItems": 64,
            "uniqueItems": true,
            "items": {
              "type": "string"
            }
          },
          "parameters": {
            "$ref": "#/components/schemas/ActionParameters"
          },
          "confirmation_token": {
            "type": "string"
          }
        },
        "required": [
          "action",
          "servers"
        ]
      },
...
    }
  }
}